refactor: impl middleware in question
