+++ internal/question/service/question_service.go
@@ -1,10 +1,9 @@
 package service
 
 import (
-	"github.com/google/uuid"
 	"github.com/tudemaha/tujuhin-be/internal/question/dto"
 )
 
 type QuestionService interface {
-	CreateQuestion(qd dto.QuestionRequestBody, owner uuid.UUID) error
+	CreateQuestion(qd dto.QuestionRequestBody, owner string) error
 }
